Code improvements
