+++ src/EphyralCorp/InputHandler.java
@@ -1,3 +1,4 @@
+package EphyralCorp;
 import java.util.Scanner;
 public class InputHandler {
